+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/).
 
 ## [Unreleased]
+
+## [18.11.2] - 2018/11/12
 ### Added
 - Subcommand for creating parallel Makefile
 
@@ -153,7 +155,8 @@ and this project adheres to [Calendar Versioning](https://calver.org/).
 ### Added
 - Everything
 
-[Unreleased]: https://github.com/opx-infra/dbp/compare/v18.11.1...HEAD
+[Unreleased]: https://github.com/opx-infra/dbp/compare/v18.11.2...HEAD
+[18.11.2]: https://github.com/opx-infra/dbp/compare/v18.11.1...v18.11.2
 [18.11.1]: https://github.com/opx-infra/dbp/compare/v18.11.0...v18.11.1
 [18.11.0]: https://github.com/opx-infra/dbp/compare/v18.10.7...v18.11.0
 [18.10.7]: https://github.com/opx-infra/dbp/compare/v18.10.6...v18.10.7